# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Narralytica API
  version: 0.1.0
  description: Canonical API contract. All shapes must reference packages/contracts/schemas/*.
servers:
  - url: http://localhost:8000
paths:
...
  /segments:
    get:
      operationId: listSegments
      parameters:
        - name: transcript_id
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List segments (optionally filtered by transcript)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Segment"
  /segments/{segment_id}:
    get:
      operationId: getSegment
      parameters:
        - name: segment_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
...
components:
  schemas:
...
    Segment:
      $ref: "./schemas/segment.schema.json"
...